data-infrastructure/streaming/requirements-kafka.txt
# Core Dependencies
aiokafka==0.8.1
confluent-kafka==2.3.0
fastavro==1.8.4
avro-python3==1.11.3

# Async/HTTP
httpx==0.27.0
websockets==12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Monitoring and Logging
prometheus-client==0.20.0
structlog>=23.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0

# Containerization
docker>=6.0.0
docker-compose>=2.20.0

# Kubernetes
kubernetes>=28.0.0

# Additional Utilities
typing-extensions>=4.5.0
backoff>=2.2.1
python-json-logger>=2.0.7
